feat: implement useMyProfileQuery for profile data management in layouts
