+++ app/layouts/settings.vue
@@ -1,47 +1,7 @@
 <script lang="ts" setup>
-import { useQuery } from '@tanstack/vue-query';
-import { meService } from '~/services/me/meService';
+import useMyProfileQuery from '~/composables/useMyProfileQuery';
 
-const userStore = useUserStore();
-
-// Helper to sync query data to store
-function syncUser(dataValue: ApiSuccessResponse<User> | undefined, err: unknown, isErr: boolean) {
-  if (!dataValue) return;
-
-  if (dataValue.success) {
-    userStore.setUser(dataValue.data);
-    userStore.error = null;
-  } else if (isErr && err) {
-    userStore.error = (err as Error).message;
-  }
-}
-
-// define query key — unique and stable
-const queryKey = ['layout-data'];
-
-// Define the query
-const { data, error, isError, suspense } = useQuery({
-  queryKey,
-  queryFn: async () => await meService.fetchProfile(),
-  // Disable re-fetch after hydration if you want to keep SSR data
-  refetchOnMount: false,
-  refetchOnWindowFocus: false,
-  staleTime: 1000 * 60 * 5, // optional: cache for 5min
-});
-
-// Reactively sync userStore when data changes
-watch(
-  () => data.value,
-  (newVal) => {
-    syncUser(newVal, error.value, isError.value);
-  },
-  { immediate: true },
-);
-
-onServerPrefetch(async () => {
-  await suspense();
-  syncUser(data.value, error.value, isError.value);
-});
+useMyProfileQuery();
 </script>
 
 <template>
